+++ app/src/main/java/uz/orifjon/educationsysteminandroid/models/Group.kt
@@ -1,3 +1,3 @@
 package uz.orifjon.educationsysteminandroid.models
 
-data class Group(var id:Long = 0,var groupName:String,var groupIsOpen:Int,var groupDate:String,var groupType:String,var courseId:Int,var mentorId:Int)
+data class Group(var id:Long = 0,var groupName:String,var groupIsOpen:Int,var groupDate:String,var groupType:String,var courseId:Long,var mentorId:Long)
